Moved Login.jsx location in directory to keep consistency with the rest of the pages structure
