+++ doggii-frontend/src/routes/AppRoutes.jsx
@@ -2,7 +2,7 @@ import { Routes, Route } from 'react-router-dom';
 import { useAuth } from '../hooks/useAuth';
 
 import ProtectedRoute from './ProtectedRoute';
-import Login from '../pages/Login';
+import Login from '../pages/Auth/Login/Login';
 import OptionsForms from '../pages/OptionsForms';
 import FormAdopter from '../pages/FormAdopter';
 import FormShelter from '../pages/FormShelter';
